Fully working calculator
We finally stomped out a mighty circular reference bug causing both leaked memory and stack overflow errors. A single line was the problem, a forgoten pop on the operator stack.

The error/warning system is working well, warnings allow solving normally but with small syntactic like a missing right parenthisis. Errors don't produce an output.

Still need to work on the command line interface and the flags we allow.